VGUI icon element
Added ir::vgui::Icon, allowing to display VMF-based icons as VGUI elements.
Added minimize and close buttons to title bar.

+++ Project/src/Iridium/rendering/model.cpp
@@ -120,7 +120,7 @@ namespace ir {
 		}
 
 		float Model::getHeight() {
-			short min { 32767 }, max { -32768 };
+			int min { 32767 }, max { -32768 };
 
 			for (auto& c : components_) {
 				for (auto& v : c) {
